+++ core/src/main/scala/org/apache/spark/deploy/history/HistoryServer.scala
@@ -262,14 +262,27 @@ class HistoryServer(
  * This launches the HistoryServer as a Spark daemon.
  */
 object HistoryServer extends Logging {
-  private val conf = new SparkConf
+  private val conf = {
+    val securityOps = ConfigSecurity.prepareEnvironment
+    val sparkConf = new SparkConf
+    if(securityOps.get("principal").isDefined
+      && securityOps.get("keytabPath").isDefined){
+      sparkConf.set("spark.history.kerberos.enabled", "true")
+        .set("spark.history.kerberos.principal", securityOps("principal"))
+        .set("spark.history.kerberos.keytab", securityOps("keytabPath"))
+    }
+    else {
+      logInfo("Security conf loaded with none Kerberos info skipping Kerberos conf")
+      sparkConf
+    }
+  }
 
   val UI_PATH_PREFIX = "/history"
 
   def main(argStrings: Array[String]): Unit = {
     try {
       Utils.initDaemon(log)
-      ConfigSecurity.prepareEnvironment
+
       new HistoryServerArguments(conf, argStrings)
       initSecurity()
       val securityManager = createSecurityManager(conf)