+++ docker/historyServer/entrypoint.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-source /root/kms_utils-0.2.1.sh
-
 source commons.sh
 
 # Create krb5.conf file
@@ -48,18 +46,9 @@ function main() {
 
    if [[ "$HDFS_KRB_ENABLE" == "true" ]]
    then
-
-     if [ ! -z "$VAULT_ROLE_ID" ]; then
-       echo "Vault role id proved, signing in"
-       login
-     fi
-     SPARK_KEYTAB_PATH="/etc/sds/spark/security"
-     getKrb userland $INSTANCE $HISTORY_SERVER_FQDN "$SPARK_KEYTAB_PATH" HISTORY_SERVER_PRINCIPAL_NAME
-
      generate_krb-conf "${KERBEROS_REALM}" "${KERBEROS_KDC_HOST}" "${KERBEROS_KADMIN_HOST}"
      mv "/tmp/krb5.conf.tmp" "/etc/krb5.conf"
-     SPARK_HISTORY_OPTS="-Dspark.history.kerberos.principal=${HISTORY_SERVER_PRINCIPAL_NAME} -Dspark.history.kerberos.keytab=${SPARK_KEYTAB_PATH}/${HISTORY_SERVER_FQDN}.keytab -Dspark.history.kerberos.enabled=true ${SPARK_HISTORY_OPTS}"
-   else
+  else
 	 echo 'HDFS SECURITY IS NOT ENABLE'
    fi